board: qti: common: Modify fdt fixup logic for early cal.
Currently, qcom,early_cal_enabled is set to "okay" by default in
kernel dts. If uboot is not able to start the early cal, it will patch
the kernel device tree to set it to "disabled".

However, when older uboot, which does not support early cal is used,
it does not patch this property and does not start the calibration,
but the kernel pcie driver thinks calibration is enabled,
times out and panics.

Modify the logic to check for qcom,early_cal_support is set to
"supported" to see if early cal is supported for the kernel uboot is
loading.

If calibration is started successfuly, patch the qcom,early_cal_enabled
property in the kernel DTS to "okay".

Change-Id: I0e87957d31c53de104d2f29bee08675b23af9d31
Signed-off-by: Vignesh Viswanathan <viswanat@qti.qualcomm.com>