mach-ipq: Add TME-IPC support for commands
This patch introduces a `comm_handler` to abstract secure operations.
This dispatcher routes calls to the appropriate communication: either
the legacy Secure Call Manager (SCM) or the TME-L IPC. This framework
is extensible and allows for adding future supports like OP-TEE.

This is necessary to support newer platforms that rely on TME for
secure functions, while maintaining backward compatibility by
defaulting to SCM for legacy boards.

This change also adds the TME-L IPC support for secure authentication
(v2/v3) and fuse reads, and refactors existing commands to use this new
modular framework.

Change-Id: I69bcd58565df5c5b6f3638aa6d998b97f6ac9663
Signed-off-by: Badhrinath S <badhrina@qti.qualcomm.com>
Signed-off-by: Timple Raj M <timple@qti.qualcomm.com>
Signed-off-by: Gurumoorthy Santhakumar <gsanthak@qti.qualcomm.com>